Added a sample testunit
@sariths I added a smaple testunit code for gridbased analysis recipe
which you can use to create your own testunits for the new classes that
you're developing. I commented the file so it should be clear how it is
structured. As far as you place the file inside tests folder
runtestunits.py should pick it up and run it.

We should commit to write the test files during the development which is
what we already do to debug the code. Let me know your thoughts.

This should partially address this issue. We still need to create a
solid test workflow for Grasshopper and Dynamo:
https://github.com/ladybug-analysis-tools/honeybee-grasshopper/issues/4