Finished modifications in datatype.py
Following up with this discussion I made a number of changes to
datatype.py file and added a unittest for datattype.

Major changes are:
1. Added a new type for RadianceValue that can be used for string
values. This type will be used for paramters such as -of.
2. The output of all of the types is wrapped in a class that has a
toRadString method that returns the string for radiance.
3. formatting style is moved to descriptors itself. for instance
isJoined for string parameters and isDualSign for Boolean values will be
defined when creating the descriptors. Check testunit for examples.

@sariths please review the code and let me know your thoughts. We can
remove string formatting class now.

This commit breaks all the development under radiance.command

I'll write the base class for RadianceParameters first and then fix
commands one by one.