Finished advancedparametersbase.py
This class can be used to create parameters from scratch using datatype.
All radiance parameters classes should inherit from this class.

@sariths you may want to take a look to usage of this code as an example
for what I was picturing for people to be able to add parameters to
commands in case they need some of the parameters that are not
implemented by us yet. They can also use static methods like
addParameterByNameAndValue or importParametersFromString but this one
benefits from all datatype developments and is more elegant.