Finished rcontrib command
This commit closes #22.

@sariths as I mentioned in #22 I just needed a very minimum
implementation of rcontrib which is basically -M for modifier. It's done
now.

I think the pattern for writing commands is coming together as:
(outputName, [input files], [specialInputs], [parameters]). I updated
gendaymtx to match this pattern and will update oconv to match the
pattern.

Also added a public method as self.checkInputFiles() which should be
called in toRadString() to address your concern about providing the
required input files before generating radiance string.

I also added a public method in commandbase as onExecution() which we
can overwrite to add specific checks for the command right before the
execution execution. .execute() calls the method before executing the
command line.

Finally I'm thinking if execute should return the output file (not the
results) on success if the command has outputFile attribute.