Added test model and started unittest for oconv #40
This commit addresses #40.
I added a test room for writing test cases that needs geometry and wrote
an example that show how to use it in action. Still need to write more
test cases and finish this test case.

@sariths FYI. I remember that you brought this up at some point about
testing with need to creating some geometry. This should be a good
starting point.