Finished HBFenSurface and started HBShadingSurface
This commit closes #42 and #43. With the minimum implementation of
HBZone now it's possible to create HBZones which can be exported
directly to radiance.

@sariths check test.py for the example that I used for testing. The
script exports all the EnegyPlus sample files to radiance.  There is
only one file that fails due to non-numeric inputs for vertices.

It handles relative system as well as absolute system. Doesn't check the
north orientation which I think should be used to rotate the sky
anyways.

The main limitation of the script is that it can only extract one
fenestration from the wall. We need a more advanced algorithm to
calculate vertices for wall with multiple openings.

Once we call the library from Grasshopper this won't be an issue since
we can use subtract and mesh the face but it would be great to have it
implemented in core honeybee library too.

Let me know your thoughts.