Modified config.py
I added a config.json file where user can set the path to installation
folders. Also cleaned the class and the method to be more generic.
Radiance bin and lib are read-only and will be set by the radiancePath.
This resolves #96. Also added a warning for #115 as we cannot really do
much in case of radiance internal commands.