Revisited radiance parameters
Improved base classes and imagebased and gridbased parameters classes.

Next step is to rename gridbased and image-based to rtrace and rpict to
follow the naming for the rest of the commands. The current push won't
break anything but the next one will break the recipes. Will push that
to master later after making all the changes.