Added SurfaceProperties and RadFile object
DO NOT MERGE THIS!

This commit has major changes which potentially break a number of
recipes because of the changes in radToString method. Here is the list
of changes.

1. Added SurfaceProperties class which includes both radiance and
energyplus properties for the surface. This is a start for creating a
unified properties class which can later be merged with OpenFOAM
properties, etc.

2. Added srfPropCollection to honeybee surfaces. Every surface should
have at list 1 SurfaceProperty for the default state. For dynamic cases
such as dynamic blinds the surface will have multiple states which can
change during the analysis by changing the state.

3. Created a RadFile object and unified radToString method between all
subclasses of AnalysisSurface. Resolved #45

4. Updated idfToRad to work with the new development. cc: @sariths