Added utilities file for daylight recipes
This helps to avoid dupliaction between method in dc and multi-phase
recipes. Also makes it easier for other devlopers to contribute and
understand the process.

@sariths I modified rfluxmtx.py and changed a number of classmethods to
staticmethods. They didn't look like classmethods to me. Also I think
rfluxmtx parameters class needs to be revised. It currently doesn't
register any of the inputs but the quality. It seems at some point you
made a decision to move them to the command itself but never updated the
parameters file. If this is the case we need to add an issue for this
and get it fixed.