Started to update image-based recipe for dc
@sariths I started to address #134 as much as I could.
The workflow is there but there are two major parts that are missing.
1. `viewSunCoeffMatrixCommands` method needs to be implemented. I don't
know how to use vwrays to calculate dc for a view in this case.
2. once we have all the images they need to be added together using
pcomb which I haven't implemented.

Once we have these two we can start debugging the whole process. I
haven't used skyMtxTotal, and skyMtxDirect which is obviously a problem!