Added window groups as static win to PIT recipes
The idea of running dynamic window groups in PIT recipes is more
confusing than what I pictured initially. It is straight forward if we
have a single window group but then if you have a number of window
groups  making decision of which combinations should be executed during
the run can be tricky. Also daylight coefficient recipe is designed just
for that.