Finished the first prototype for imagebased dc
Now we can handle result visualization for multiple windows groups with
multiple states! Now implementing 5-phase and 3-Phase image-based stuff
is just a matter of implementing the recipe for the commands which we
already have! I think in this case having 5-phase implemented for
image-based makes a lot of sense.

Resolved #134