manual fixes
I checked the files one by one and fixed the issues. Most of the issues
are the result of poor find and replace method that I chose. I should
have found a list of all class names and then implement extra checks
only for methods and variables with the same name as a class. otherwise
the update was very good.