Finished adding materials
I re-structured materials. Now custom material is removed and RadianceMaterial class is minimized. The methods and properties of Custom class is now merged into Primitive class.
This change has made everything cleaner and also made it possible to have a place holder for all the radiance material types with minimum amount of code.