Use python's default parameters instead of explictly setting variables
There's no need to explicitly set variables to what they've been
declared as default function parameters. This can also be seen when
defaulting a parameter value to `None`, and subsequently setting the
value in the function body explicitly.  It's at best dead code which
needlessly requires the programmer to keep track of additional details
when making changes.

At worst, there may be unintended consequences if the code is written
carelessly. For example, in
```
    def __init__(self, illuminance_value=10000, sky_type=0, suffix=None):
        self.illuminance_value = illuminance_value or 10000
```
if we try to instantiate `foo(0)`, this instance would have an
illuminance value of 10000 instead of 0.

Both CPython and IronPython honor the default parameters, so there's no
reason to set them explicitly.